Moved APIService struct & handler methods to handler package
